Paste image content from the clipboard
This allows someone to take a screen capture and send it to a peer without having to save it as a file.  This only works if only one peer is connected as I wasn't sure what the UI flow should be if there are multiple peers.  If nothing else, maybe this can be used to implement the full flow for multiple peers. =]