No Webapps at the moment
Is there a reason why "apple-mobile-web-app-capable" is set to "no"? This results in the whole "<!-- Web App Config -->" section being unnecessary since there is no web-app and snapdrop is opened in Safari all the time. Therefore I would suggest to change this value in order to give a better experience. If this would be switched to "yes" I have also added the "apple-mobile-web-app-status-bar-style" so the status bar will conform to the app-UI.