feat(client): specified more info for the user
I added "with the same network" because the first time that I used this webapp I did't thought to connect my phone to the same network of my pc, obviously I fast understood the problem and connected the phone to the same network. But for a basic user this can't be easy to understand. I made this little fix just for the UX.
Sincerely, D.